requirements.txt
backcall==0.1.0
cachelib==0.1
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
colorama==0.4.3
dump==0.0.5
Flask==1.1.2
Flask-Session==0.3.2
Flask-SQLAlchemy==2.4.3
get==2019.4.13
helpers==0.2.0
itsdangerous==1.1.0
jedi==0.17.0
Jinja2==2.11.2
MarkupSafe==1.1.1
parso==0.7.0
passlib==1.7.2
post==2019.4.13
prompt-toolkit==3.0.5
psycopg2==2.8.5
public==2019.4.13
Pygments==2.6.1
python-dotenv==0.13.0
query-string==2019.4.13
request==2019.4.13
requests==2.23.0
SQLAlchemy==1.3.17
tornado==6.0.4
traitlets==4.3.3
urllib3==1.25.9
wcwidth==0.1.9
Werkzeug==1.0.1
wincertstore==0.2
gunicorn==19.9.0